+++ src/entities/post.py
@@ -1,4 +1,5 @@
 class Diary:
-    def __init__(self, post, user=None):
-        self.post = post
-        self.user = user
+    def __init__(self, username, message, created_at):
+        self.username = username
+        self.message = message
+        self.created_at = created_at